requirements.txt
# Project dependencies (agent uses PostgreSQL via psycopg)
psycopg[binary]==3.3.*
<<<<<<< Updated upstream
SQLAlchemy>=2.0.0
=======
openai-agents
>>>>>>> Stashed changes
python-dotenv>=1.0.0
grpcio>=1.78.0
grpcio-tools>=1.78.0
psutil>=6.0.0
